Persist scan settings in archives
